fix: 修复wechat_official_service类型错误，优化CI构建配置
- 修复 wechat_official_service.dart 中 expiresIn 类型错误 (num -> int)
- iOS 构建工作流：flutter analyze 添加 --no-fatal-infos --no-fatal-warnings
- CI 工作流：同样优化 analyze 配置，避免 warning 导致构建失败

+++ lib/services/wechat_official_service.dart
@@ -82,7 +82,7 @@ class WechatOfficialService {
 
       final data = jsonDecode(resp.body);
       final token = data['access_token'];
-      final expiresIn = data['expires_in'] ?? 7200;
+      final expiresIn = (data['expires_in'] ?? 7200) as int;
 
       if (token is String && token.isNotEmpty) {
         final prefs = await _sp;
